refactor(provider): relocate config resolver to agent-core + SSOT cost fields (ARCH-PROVIDER-003)
Step 1/7. normalizeProviderConfig + createProviderFromConfig move from agent-executor
into agent-core (they import agent-core symbols only), so the coming dag-node-llm-text
leaf can reuse the one credential-resolution path while depending on agent-core alone.
agent-executor re-exports both; profile helpers (process.env-reading) stay put.

Additive SSOT fields on IProviderDefinition: allowedModels (enforced model allowlist,
distinct from the descriptive modelCatalog) and per-model costPerInputToken/
costPerOutputToken on IProviderModelCatalogEntry (cost is a per-model attribute).
